+++ server/src/server.ts
@@ -3,11 +3,10 @@ import {
   createConnection,
   InitializeParams,
   ProposedFeatures,
-  TextDocumentPositionParams,
   TextDocuments,
 } from 'vscode-languageserver'
 
-import { validateTextDocument, completition } from './linter';
+import { validateTextDocument, completions } from './linter';
 import { initializeSettings, settingsChanged } from './settings'
 
 // Create a connection for the server. The connection uses Node's IPC as a transport.
@@ -22,9 +21,9 @@ connection.onInitialize((params: InitializeParams) => {
   initializeSettings(connection, params.capabilities)
 
   return {
+    // Tell the client which capabilities the server supports
     capabilities: {
       textDocumentSync: documents.syncKind,
-      // Tell the client that the server supports code completion
       completionProvider: {
         resolveProvider: true
       }
@@ -51,26 +50,15 @@ connection.onDidChangeWatchedFiles(_change => {
 })
 
 // This handler provides the initial list of the completion items.
-connection.onCompletion(
-  (_textDocumentPosition: TextDocumentPositionParams): CompletionItem[] => {
-    // The pass parameter contains the position of the text document in
-    // which code complete got requested. For the example we ignore this
-    // info and always provide the same completion items.
-    return completition(_textDocumentPosition)
-  }
-)
+connection.onCompletion(completions)
 
-// This handler resolves additional information for the item selected in
-// the completion list.
+// This handler resolves additional information for the item selected in the completion list.
 connection.onCompletionResolve(
   (item: CompletionItem): CompletionItem => {
-    if (item.data === 1) {
-      item.detail = 'TypeScript details'
-      item.documentation = 'TypeScript documentation'
-    } else if (item.data === 2) {
-      item.detail = 'JavaScript details'
-      item.documentation = 'JavaScript documentation'
-    }
+    // if (item.data === 1) {
+    //   item.detail = 'TypeScript details'
+    //   item.documentation = 'TypeScript documentation'
+    // }
     return item
   }
 )
